Backend/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
attrs==24.3.0
bcrypt==5.0.0
beautifulsoup4==4.13.3
blinker==1.9.0
boto3==1.36.3
botocore==1.36.3
build==1.4.0
certifi==2024.12.14
cffi==2.0.0
charset-normalizer==3.4.1
chromadb==1.5.5
click==8.3.1
cmake==3.31.4
colorama==0.4.6
contourpy==1.3.1
cryptography==46.0.5
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
durationpy==0.10
et_xmlfile==2.0.0
filelock==3.17.0
filetype==1.2.0
fire==0.7.0
Flask==3.1.3
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
flatbuffers==25.1.21
fonttools==4.55.4
fpdf2==2.8.7
frozenlist==1.8.0
fsspec==2024.12.0
gast==0.6.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.0
google-auth-httplib2==0.3.0
google-genai==1.67.0
google-generativeai==0.8.6
google-pasta==0.2.0
googleapis-common-protos==1.73.0
greenlet==3.3.0
grpcio==1.78.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
h5py==3.12.1
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.24.7
idna==3.7
importlib_metadata==8.7.1
importlib_resources==6.5.2
itsdangerous==2.2.0
jax==0.5.0
jaxlib==0.5.0
Jinja2==3.1.5
jiter==0.13.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==3.8.0
kiwisolver==1.4.8
kubernetes==35.0.0
langchain==1.2.11
langchain-classic==1.0.2
langchain-community==0.4.1
langchain-core==1.2.18
langchain-openai==1.1.11
langchain-text-splitters==1.1.1
langgraph==1.1.0
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.11
langsmith==0.7.16
libclang==18.1.1
lxml==6.0.2
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.2
matplotlib==3.10.0
mdurl==0.1.2
mediapipe==0.10.20
ml-dtypes==0.4.1
mmh3==5.2.1
mpmath==1.3.0
mtcnn==1.0.0
multidict==6.7.1
mypy_extensions==1.1.0
namex==0.0.8
networkx==3.4.2
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.24.3
openai==2.26.0
opencv-contrib-python==4.11.0.86
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opt_einsum==3.4.0
optree==0.14.0
orjson==3.11.7
ormsgpack==1.12.2
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psutil==6.1.1
py-cpuinfo==9.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybase64==1.4.3
pybboxes==0.1.6
pycparser==2.22
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.1
pyparsing==3.2.1
PyPika==0.51.1
pyproject_hooks==1.2.0
PySocks==1.7.1
pythainlp==5.3.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.37.0
regex==2026.2.28
reportlab==4.4.10
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
roboflow==1.1.51
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.11.1
sahi==0.11.20
scikit-learn==1.8.0
scipy==1.15.1
seaborn==0.13.2
sentencepiece==0.2.0
setuptools==75.8.0
shapely==2.0.6
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
SQLAlchemy==2.0.45
sympy==1.13.1
tenacity==9.1.4
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.1
ultralytics==8.3.65
ultralytics-thop==2.0.14
uritemplate==4.2.0
urllib3==2.3.0
uuid_utils==0.14.1
uvicorn==0.41.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.23.0
yolov5==7.0.14
zipp==3.23.0
zstandard==0.25.0
